Altera estrutura de pastas para uma melhor organização do projeto.
