--- src/main/java/AlteracaoPreco.java
@@ -1,6 +1,0 @@
-import java.math.BigDecimal;
-
-public interface AlteracaoPreco {
-    BigDecimal alteraPreco(BigDecimal preco);
-
-}
