--- src/main/java/CalculaPreco.java
@@ -1,5 +1,0 @@
-import java.math.BigDecimal;
-
-public interface CalculaPreco<T> {
-    BigDecimal calcular(T t);
-}
