--- src/main/java/Frete.java
@@ -1,9 +1,0 @@
-import java.math.BigDecimal;
-
-public class Frete implements AlteracaoPreco{
-    @Override
-    public BigDecimal alteraPreco(BigDecimal preco) {
-
-        return preco.multiply(BigDecimal.valueOf(0.1));
-    }
-}
