--- src/main/java/PessoaFisica.java
@@ -1,11 +1,0 @@
-public class PessoaFisica extends Pessoa{
-    private String cpf;
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-}
