--- src/main/java/PessoaJuridica.java
@@ -1,11 +1,0 @@
-public class PessoaJuridica extends Pessoa{
-    String cnpj;
-
-    public String getCnpj() {
-        return cnpj;
-    }
-
-    public void setCnpj(String cnpj) {
-        this.cnpj = cnpj;
-    }
-}
