--- src/main/java/Promocao.java
@@ -1,10 +1,0 @@
-import java.math.BigDecimal;
-
-public class Promocao implements AlteracaoPreco {
-
-    @Override
-    public BigDecimal alteraPreco(BigDecimal preco) {
-
-        return preco.multiply(BigDecimal.valueOf(-0.15));
-    }
-}
