+++ src/main/java/model/Pessoa.java
@@ -1,0 +1,33 @@
+package model;
+
+public abstract class Pessoa {
+    private String nome;
+    private String email;
+    private String telefone;
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getTelefone() {
+        return telefone;
+    }
+
+    public void setTelefone(String telefone) {
+        this.telefone = telefone;
+    }
+
+
+}
