+++ src/main/java/model/PessoaFisica.java
@@ -1,0 +1,13 @@
+package model;
+
+public class PessoaFisica extends Pessoa{
+    private String cpf;
+
+    public String getCpf() {
+        return cpf;
+    }
+
+    public void setCpf(String cpf) {
+        this.cpf = cpf;
+    }
+}
