+++ src/main/java/model/PessoaJuridica.java
@@ -1,0 +1,13 @@
+package model;
+
+public class PessoaJuridica extends Pessoa{
+    String cnpj;
+
+    public String getCnpj() {
+        return cnpj;
+    }
+
+    public void setCnpj(String cnpj) {
+        this.cnpj = cnpj;
+    }
+}
