+++ src/main/java/repository/ProdutosPadraoArquivo.java
@@ -1,15 +1,12 @@
+package repository;
+
+import model.Produto;
 import repository.RepositorioBase;
 
-import java.io.IOException;
-import java.nio.channels.Channels;
-import java.nio.channels.FileChannel;
-import java.nio.channels.FileLock;
 import java.nio.file.Path;
-import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Properties;
 
 public class ProdutosPadraoArquivo implements RepositorioBase<Produto, Integer> {
