+++ src/main/java/service/AlteracaoPreco.java
@@ -1,0 +1,8 @@
+package service;
+
+import java.math.BigDecimal;
+
+public interface AlteracaoPreco {
+    BigDecimal alteraPreco(BigDecimal preco);
+
+}
