+++ src/main/java/service/CalculaPreco.java
@@ -1,0 +1,7 @@
+package service;
+
+import java.math.BigDecimal;
+
+public interface CalculaPreco<T> {
+    BigDecimal calcular(T t);
+}
