+++ src/main/java/service/ListaProdutos.java
@@ -1,3 +1,7 @@
+package service;
+
+import repository.Carrinho;
+
 public interface ListaProdutos {
 
     void listarProdutos(Carrinho carrinho);
